Refactor code and add validation to input fields
