+++ frontend/src/components/Loading.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
 const Loading = () => (
-    <div className="flex flex-col items-center w-full h-screen">
-        <h2 className="MainHeading">Loading...</h2>
-    </div>
+    <section className="SectionHolder">
+        <div className="flex flex-col items-center w-full h-screen">
+            <h2 className="MainHeading">Loading...</h2>
+        </div>
+    </section>
 );
 
 export default Loading;
