+++ frontend/src/pages/home/Home.jsx
@@ -3,15 +3,12 @@ import { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { fetchStatesList } from '../../redux/states/states';
 import Loading from '../../components/Loading';
-import Button from '../../components/inputs/Button';
 
-const TaxForm = React.lazy(() => import('../../components/Form'));
-const TaxList = React.lazy(() => import('../../components/TaxList'));
+const HomeMainSection = React.lazy(() => import('../../components/HomeMainSection'));
 
 const Home = () => {
     const dispatch = useDispatch();
     const { loading, stateList } = useSelector((state) => state.states);
-    const { newTaxData } = useSelector((state) => state.taxes);
 
     useEffect(() => {
         if (stateList.length === 0) {
@@ -19,33 +16,16 @@ const Home = () => {
         }
     }, [dispatch]);
 
-    const SyncWithDB = () => {
-        console.log('Sync', newTaxData);
-    }
-
     return (
         <Suspense fallback={<Loading />}>
             <div id="Home-holder">
-                <section className="SectionHolder">
-                    {loading && (<Loading />)}
-                    {(!loading && stateList.length > 0) && (
-                        <>
-                            <div className="flex flex-col w-full">
-                                <Button
-                                    type='button' className='ButtonRight'
-                                    handleClick={SyncWithDB} label='Sync With DB'
-                                />
-                            </div>
-                            <TaxForm />
-                        </>
+                {loading && (<Loading />)}
 
-                    )}
-                </section>
-                <section className="SectionHolder">
-                    <TaxList />
-                </section>
+                {(!loading && stateList.length > 0) && (
+                    <HomeMainSection />
+                )}
             </div>
-        </Suspense>
+        </Suspense >
     );
 }
